Versionner le schéma avec Flyway (ddl-auto: validate)
Le schéma n'est plus déduit des entités au démarrage : ddl-auto: update est
pratique en développement mais ne supprime ni ne renomme rien, n'est pas
reproductible et laisse le schéma diverger en silence.

- V1__init_schema.sql : schéma initial, dérivé du schéma réellement produit
  par Hibernate puis nettoyé (contraintes nommées, commentaires, index
  adaptés aux requêtes de l'API)
- ddl-auto: validate — toute divergence entité/schéma fait échouer le
  démarrage au lieu de passer inaperçue
- Tests inchangés : H2 en mémoire, Flyway désactivé
- Flyway 10.20.1 fixé explicitement (la version gérée par Boot 3.3.0 diffère)

Vérifié sur base vierge pour les 5 projets : Flyway crée le schéma et
Hibernate le valide.

+++ src/main/java/com/kfokam48/dto/ConvertRequest.java
@@ -8,6 +8,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 /**
  * Requête de conversion : devise source, devise cible et montant.
  *
@@ -27,7 +29,12 @@ public class ConvertRequest {
     @Pattern(regexp = "^[A-Za-z]{3}$", message = "Target currency must be a 3-letter ISO code")
     private String toCurrency;
 
+    /**
+     * Montant à convertir. BigDecimal et non double : un montant monétaire ne
+     * doit jamais transiter par une virgule flottante binaire, qui ne peut pas
+     * représenter exactement des valeurs comme 0,10 et accumule des écarts.
+     */
     @NotNull(message = "Amount is required")
     @Positive(message = "Amount must be positive")
-    private Double amount;
+    private BigDecimal amount;
 }
